Changed landing page and set dark theme shortcuts
