+++ assets/translations/en.json
@@ -26,9 +26,6 @@
     "terms_and_agreement": "I agree with &nbsp;<a href=\"https://legacy.dooboolab.com/termsofservice\">Terms for Agreement</a>&nbsp;and&nbsp;<a href=\"https://legacy.dooboolab.com/privacyandpolicy\">Privacy and Policy</a>.",
     "missing_inputs": "Some information is missing. Please fill all information."
   },
-  "Intro": {
-    "welcome": "Hello there ~ 🙌 Welcome to WeCount."
-  },
   "Onboarding": {
     "hello": "Hellow there. 🙌",
     "welcome": "Welcome to Wecount.",
