+++ assets/translations/ko.json
@@ -28,9 +28,6 @@
     "terms_and_agreement": "위카운트 <a href=\"https://legacy.dooboolab.com/termsofservice\">이용약관</a> 및 <a href=\"https://legacy.dooboolab.com/privacyandpolicy\">개인정보취급방침</a>에 모두 동의합니다.",
     "missing_inputs": "모든 항목이 기재되지 않았습니다. 회원가입을 위해 모든 정보를 입력해주세요."
   },
-  "Intro": {
-    "welcome": "안녕하세요. 🙌 WE카운트에 오신 것을 환영합니다."
-  },
   "Onboarding": {
     "hello": "안녕하세요. 🙌",
     "welcome": "WE카운트에 오신 것을 환영합니다.",
