+++ src/components/uis/header/Manipulation.svelte
@@ -14,12 +14,29 @@
       display: none;
     }
   }
+  .profile {
+    cursor: pointer;
+  }
 </style>
 
-<script lnag="ts">
+<script lang="ts">
+  import {push, replace} from 'svelte-spa-router';
+  import supabase from '../../../lib/db';
+  import {user} from '../../../stores/sessionStore';
   import {SvgBell} from '../../../utils/Icon';
   import Profile from './Profile.svelte';
   import Search from './Serach.svelte';
+
+  const signInOut = async (): Promise<void> => {
+    if ($user) {
+      await supabase.auth.signOut();
+      await replace('/');
+
+      return;
+    }
+
+    await push('/sign_in');
+  };
 </script>
 
 <div class="manipulation">
@@ -29,7 +46,7 @@
   <div class="notification">
     <SvgBell />
   </div>
-  <div>
+  <div class="profile" on:click={signInOut}>
     <Profile name="hanna" imageSrc={'https://picsum.photos/200'} />
   </div>
 </div>
